Added confest file with generate_university
