+++ pytest.ini
@@ -1,3 +1,3 @@
 [pytest]
 DJANGO_SETTINGS_MODULE = AcadeMe.settings
-python_files = tests.py test_*.py *_tests.py
+python_files = tests.py test_*.py *_tests.py conftest.py
